+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-DC0cdsXh.js
@@ -1,8 +1,0 @@
-import{c as r,s as e}from"./flowDiagram-HODETNUW-VZUDdT29.js";import{_ as a}from"./mermaid.core-Du88hnd2.js";import"./chunk-5VM5RSS4-Ca2tShVP.js";import"./chunk-XXDRQBXY-DVvHJ5op.js";import"./chunk-POPQ4Y6H-YsCCqkcH.js";import"./chunk-F27PBJKO-CYhB_AKq.js";import"./channel-CvZZ0sM1.js";import"./index-frx2-TkG.js";var o=a(t=>`${e(t)}
-  .swimlane.cluster rect {
-    stroke: ${t.clusterBorder} !important;
-  }
-  [data-look="neo"].cluster rect {
-    filter: none;
-  }
-`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};