+++ internal/web/public/assets/swimlanesDiagram-VR7AAH4N-bdy9LLyJ.js
@@ -1,0 +1,8 @@
+import{c as r,s as e}from"./flowDiagram-HODETNUW-Bjt_Wdw4.js";import{_ as a}from"./mermaid.core-qsbcLsxR.js";import"./chunk-5VM5RSS4-B1LQXnwr.js";import"./chunk-XXDRQBXY-CaGX4A7W.js";import"./chunk-POPQ4Y6H-Bu-kO8U4.js";import"./chunk-F27PBJKO-CuOok2Ub.js";import"./channel-BmZg-JAk.js";import"./index-BAQL9Aat.js";var o=a(t=>`${e(t)}
+  .swimlane.cluster rect {
+    stroke: ${t.clusterBorder} !important;
+  }
+  [data-look="neo"].cluster rect {
+    filter: none;
+  }
+`,"getStyles"),s=o,n=r({defaultLayout:"swimlane",styles:s});export{n as diagram};